% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covSS.R
\name{covSS}
\alias{covSS}
\title{covSS}
\usage{
covSS(
  data,
  genoMeans = NULL,
  reps = NULL,
  use_nearPD = TRUE,
  sepExp = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Dataframe containing the genotypes in the first column. The final column
must contain the focal trait (if there is any).}

\item{genoMeans}{Genotypic means. \code{NULL} by default and calculated internally. Can be specified to avoid
repeated calculations if covSS is called for the same genotypes multiple times.}

\item{reps}{Number of replicates. Also \code{NULL} by default but can be specified for the same reasons
as \code{genoMeans}.}

\item{use_nearPD}{Boolean indicating whether \code{Matrix::nearPD} should be used to enforce PDness of the
genetic covariance matrix.}

\item{sepExp}{Boolean specifying whether the focal trait was measured on different plants than the secondary
features. Only use if there is a focal trait. Sets the residual covariances between secondary features or factors
and the focal to \code{0}.}

\item{verbose}{Boolean}
}
\value{
A list containing the phenotypic, genetic, and residual covariance matrices.
}
\description{
\code{covSS} calculates the phenotypic, genetic, and residual covariance matrix.
}
