% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolds.R
\name{createFolds}
\alias{createFolds}
\title{createFolds}
\usage{
createFolds(genos, folds = 5)
}
\arguments{
\item{genos}{Character vector containing the genotypes.}

\item{folds}{Number of folds to be created}
}
\value{
A list of vectors with the genotypes for each fold.
}
\description{
\code{createFolds} creates folds out of a list of genotypes.
}
