% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModel_slow.R
\name{factorModel_slow}
\alias{factorModel_slow}
\title{factorModel_slow}
\usage{
factorModel_slow(
  data,
  cormat,
  m = NULL,
  lower = 0.1,
  what = "genetic",
  verbose = TRUE
)
}
\arguments{
\item{data}{The datamatrix used to estimate the (regularized) correlation matrix used to fit the factor model. Only used to determine the number of
individuals (in case of a phenotypic or residual correlation matrix) or number of genotypes (in case of a genetic correlation matrix).}

\item{cormat}{The correlation matrix.}

\item{m}{Latent dimension.}

\item{lower}{Lower bound for the uniquenesses.}

\item{what}{What type of correlation matrix will be used to fit the factor model.}

\item{verbose}{Boolean}
}
\value{
A list with the varimax rotated loadings, uniquenesses, rotation matrix, and latent dimension.
}
\description{
\code{factorModel_slow} fits the ML factor model using a correlation matrix and a specified latent dimension.
}
\keyword{internal}
