% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorScores.R
\name{factorScores}
\alias{factorScores}
\title{factorScores}
\usage{
factorScores(data, loadings, uniquenesses, m, type = "thomson")
}
\arguments{
\item{data}{Dataframe containing the secondary features used to estimate the factor scores. Genotypes should be in the first column.}

\item{loadings}{Loadings matrix.}

\item{uniquenesses}{Uniquenesses matrix.}

\item{m}{Latent dimension.}

\item{type}{Type of factor scores to be calculated.}
}
\value{
A dataframe with the factor scores.
}
\description{
\code{factorScores} calculates the factor scores for each individual based on the loadings and uniquenesses, as well as the latent dimension.
}
