% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorSelect.R
\name{factorSelect}
\alias{factorSelect}
\title{factorSelect}
\usage{
factorSelect(data, procedure = "leaps", verbose = TRUE)
}
\arguments{
\item{data}{A dataframe containing the genotypes in the first column, followed by the factors, and finally the focal trait.}

\item{procedure}{Only \code{"leaps"} for now.}

\item{verbose}{Boolean}
}
\value{
A vector with the names of the selected factors.
}
\description{
\code{factorSelect} uses the specified procedure to determine which factors should be used in genomic prediction.
}
