% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBLUPo.R
\name{gBLUPo}
\alias{gBLUPo}
\title{gBLUPo}
\usage{
gBLUPo(
  Y,
  K,
  X = matrix(rep(1, nrow(K[train.set, train.set]))),
  B = t(matrix(rep(0, ncol(Y)))),
  Sg,
  Se,
  train.set
)
}
\arguments{
\item{Y}{A matrix containing the mean factor and focal trait values. MUST NOT CONTAIN THE GENOTYPE COLUMN.}

\item{K}{The kinship matrix.}

\item{X}{Matrix of covariates.}

\item{B}{Matrix of coefficients.}

\item{Sg}{Possibly subsetted genetic covariance matrix of factors and focal trait.}

\item{Se}{Possibly subsetted residual covariance matrix of factors and focal trait.}

\item{train.set}{Character vector that contains the names of the training set genotypes.}
}
\value{
A list containing the training set gBLUPs for all factors/traits, or just the focal trait.
}
\description{
A fast implementation to calculate the training set gBLUPs.
}
\keyword{internal}
