% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoMeans.R
\name{genoMeans}
\alias{genoMeans}
\title{genoMeans}
\usage{
genoMeans(data)
}
\arguments{
\item{data}{Dataframe containing the genotypes in the first column.}
}
\value{
A dataframe with genotypic means.
}
\description{
\code{.genoMeans} calculates the per-genotype means for secondary or focal trait data.
}
\keyword{internal}
