% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeMeans.R
\name{genotypeMeans}
\alias{genotypeMeans}
\title{genotypeMeans}
\usage{
genotypeMeans(data)
}
\arguments{
\item{data}{Dataframe containing the genotypes in the first column.}
}
\value{
A dataframe with genotypic means.
}
\description{
\code{genotypeMeans} calculates the per-genotype means for secondary or focal trait data.
}
\keyword{internal}
