% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFAsim.R
\name{GFAsim}
\alias{GFAsim}
\title{GFAsim}
\usage{
GFAsim(
  K,
  r,
  n.LSF,
  n.LNF,
  LSF.rho = 0,
  LNF.rho = 0,
  LSNF.rho = 0,
  S.per.LF = 10,
  Y.psi,
  L.min = 0.3,
  L.max = 0.8,
  S.sg2,
  S.se2,
  Y.sg2,
  Y.se2,
  resCors = FALSE
)
}
\arguments{
\item{K}{Kinship matrix.}

\item{r}{Number of replicates per genotype.}

\item{n.LSF}{Number of latent signal factors (i.e. factors that the focal trait loads on).}

\item{n.LNF}{Number of latent noise factors.}

\item{LSF.rho}{Between latent signal factor correlation.}

\item{LNF.rho}{Between latent noise factor correlation.}

\item{LSNF.rho}{Correlation between latent signal and noise factors.}

\item{S.per.LF}{Number of observed features per latent factor.}

\item{Y.psi}{Focal trait unique variance (i.e. 1 minus the communality).}

\item{L.min}{Minimum absolute loading for the observed features.}

\item{L.max}{Maximum absolute loading for the observed features.}

\item{S.sg2}{Observed feature genetic variance.}

\item{S.se2}{Observed feature residual variance.}

\item{Y.sg2}{Focal trait genetic variance.}

\item{Y.se2}{Focal trait residual variance.}

\item{resCors}{Should there be random residual correlations among secondary features
and between sec. features and the focal trait? Residual correlations will be normally
distributed with mean 0 and a standard deviation of approximately 0.05.}
}
\value{
A list with the "real" simulated dataframe, benchmark dataframe, and the corresponding genetic and residual covariance matrices.
The list also contains the prediction target and simulated BLUEs for all features.
}
\description{
\code{GFAsim} simulates genetic latent factor driven data, including the actual latent factor scores for benchmarking purposes.
}
