% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfBLUP.R
\name{gfBLUP}
\alias{gfBLUP}
\title{gfBLUP}
\usage{
gfBLUP(data, selection = NULL, K, sepExp = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{The dataframe containing the secondary features (factors) and the focal trait. The column
order should be G-F1-...-Fm-Y.}

\item{selection}{Optional argument. Output from \code{gfBLUP::factorSelect}. All factors are used if left unspecified.}

\item{K}{The kinship matrix with the genotype identifiers as row and column names.}

\item{sepExp}{Boolean indicating whether the focal trait and secondary features were measured on different plants/plots.}

\item{verbose}{Boolean}
}
\value{
A list with the test set predictions, factor selection, and the genetic and residual covariance matrices of factors
and focal trait, as well as their plot-level heritabilities.
}
\description{
\code{gfBLUP} uses genetic factors to produce multi-trait genomic predictions.
}
