% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixRoot.R
\name{MatrixRoot}
\alias{MatrixRoot}
\title{MatrixRoot}
\usage{
MatrixRoot(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
The root of the matrix \code{x}.
}
\description{
Calculates the root of a matrix.
}
\keyword{internal}
