% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precalcVecs.R
\name{precalcVecs}
\alias{precalcVecs}
\title{precalcVecs}
\usage{
precalcVecs(data, genoMeans, reps, what, folds, n.folds, verbose, dopar)
}
\arguments{
\item{data}{A dataframe containing the genotypes in the first column and secondary features
in the other columns.}

\item{genoMeans}{A dataframe containing the genotypic means, with the genotypes in the first column.}

\item{reps}{Number of replicates}

\item{what}{Pre-calculation of vectors for the genetic or residual correlation matrix?}

\item{folds}{List of vectors with the genotypes in the folds}

\item{n.folds}{Number of folds}

\item{verbose}{Verbosity}

\item{dopar}{Parallel calculation or not}
}
\value{
A list of lists containing the required vectors for the regularization
}
\description{
precalcVecs is an internal function that pre-calculates the vectors (possibly in parallel) used
in the efficient regularization of the phenotypic/genetic/residual correlation matrix.
}
\keyword{internal}
