% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancyFilter.R
\name{redundancyFilter}
\alias{redundancyFilter}
\title{redundancyFilter}
\usage{
redundancyFilter(data, tau = 0.95, verbose = TRUE)
}
\arguments{
\item{data}{Data matrix to be filtered based on genetic correlations. First column contains genotype identifiers.}

\item{tau}{Threshold value. The filtered matrix will have no pair of features with an absolute
correlation larger than this threshold. Defaults to \code{0.95}.}

\item{verbose}{Logical enabling verbose mode.}
}
\value{
A list containing the filtered correlation matrix and filtered secondary feature matrix.
}
\description{
\code{redundancyFilter} is a wrapper around \code{FMradio::RF} that redundancy filters a correlation matrix
given a specified correlation threshold.
}
