% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCor_corlw.R
\name{regCor_corlw}
\alias{regCor_corlw}
\title{regCor_corlw}
\usage{
regCor_corlw(R, penalty)
}
\arguments{
\item{R}{A correlation matrix.}

\item{penalty}{The regularization penalty.}
}
\value{
A penalized correlation matrix.
}
\description{
\code{.regCor_corlw} calculates the penalized correlation matrix using an identity
matrix target.
}
\keyword{internal}
