% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCor_kcvl.R
\name{regCor_kcvl}
\alias{regCor_kcvl}
\title{regCor_kcvl}
\usage{
regCor_kcvl(penalty, data, folds, genoMeans, reps, dopar, use_nearPD, what)
}
\arguments{
\item{penalty}{Defaults to \code{NULL} but can be specified to override the data-driven regularization.}

\item{data}{A dataframe containing the genotypes in the first column and the secondary features in the other columns.}

\item{folds}{Either a numeric value specifying the number of folds that should be used or a list of vectors each containing the genotypes for each fold.}

\item{genoMeans}{Dataframe with genotypic means.}

\item{reps}{Number of replicates.}

\item{dopar}{Boolean specifying whether to use parallelization.}

\item{use_nearPD}{Should \code{Matrix::nearPD} be used to enforce PDness?}

\item{what}{Type of correlation matrix that will be regularized.}
}
\value{
The cross-validated negative log-likelihood to be minimized.
}
\description{
\code{.regCor_kcvl} provides the function that should be optimized to find the optimal regularization penalty.
}
\keyword{internal}
