% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCor_ll.R
\name{regCor_ll}
\alias{regCor_ll}
\title{regCor_ll}
\usage{
regCor_ll(S, R)
}
\arguments{
\item{S}{A correlation matrix based on the left out fold.}

\item{R}{A correlation matrix based on the other folds.}
}
\value{
The negative log-likelihood.
}
\description{
\code{.regCor_ll} calculates the negative log-likelihood assuming an identity
matrix regularization target.
}
\keyword{internal}
