% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizedCorrelation_2.R
\name{regularizedCorrelation_2}
\alias{regularizedCorrelation_2}
\title{regularizedCorrelation_2}
\usage{
regularizedCorrelation_2(
  data,
  folds = 5,
  what = "genetic",
  dopar = FALSE,
  penalty = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{A dataframe containing the genotypes in the first column and secondary features
in the other columns.}

\item{folds}{A list of vectors containing the genotypes for each fold, or a number specifying how
many folds must be created.}

\item{what}{Which correlation matrix must be regularized? Can be \code{"genetic"} or \code{"residual"}.}

\item{dopar}{Boolean specifying whether to use parallelization or not.}

\item{penalty}{An optional argument containing a user-specified penalty between \code{0} and \code{1}. Specifying one overrides the data-driven regularization.}

\item{verbose}{Boolean.}
}
\value{
A list containing the optimal penalty and the regularized correlation matrix.
}
\description{
\code{regularizedCorrelation_2} regularizes a phenotypic, genetic, or residual correlation matrix
using k-fold cross-validation.
}
