% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMVData.R
\name{simMVData}
\alias{simMVData}
\title{simMVData}
\usage{
simMVData(rowCov, colCov)
}
\arguments{
\item{rowCov}{Row covariance matrix (e.g. kinship or identity).}

\item{colCov}{Column covariance matrix (e.g. genetic covariance matrix).}
}
\value{
A data matrix.
}
\description{
\code{simMVData} simulates multi-trait data using given row and column covariance matrices.
}
\keyword{internal}
