% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\title{subset}
\usage{
subset(data, filter)
}
\arguments{
\item{data}{A dataframe containing the features to be filtered. The first column contains the genotypes, the last
column contains the focal trait.}

\item{filter}{The redundancy filtered correlation matrix containing the features that should be kept.}
}
\value{
A redundancy filtered dataframe.
}
\description{
\code{subset} is a wrapper around \code{FMradio::subSet} that subsets a dataframe using
a redundancy filtered correlation matrix.
}
