% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetRF.R
\name{subsetRF}
\alias{subsetRF}
\title{subsetRF}
\usage{
subsetRF(sec, filter)
}
\arguments{
\item{sec}{A dataframe containing the secondary features to be filtered. The first column contains the genotypes.}

\item{filter}{The redundancy filtered correlation matrix containing the features that should be kept.}
}
\value{
A redundancy filtered secondary feature dataframe.
}
\description{
\code{subsetRF} is a wrapper around \code{FMradio::subSet} that subsets a secondary feature dataframe using
a redundancy filtered correlation matrix.
}
