% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train2test.R
\name{train2test}
\alias{train2test}
\title{train2test}
\usage{
train2test(BLUPs, K, train.set, test.set)
}
\arguments{
\item{BLUPs}{A vector containing the training set focal trait BLUPs.}

\item{K}{The kinship matrix.}

\item{train.set}{A character vector containing the names of the training set genotypes.}

\item{test.set}{A character vector containing the names of the test set genotypes.}
}
\value{
A vector containing the test set BLUPs.
}
\description{
A function to convert training set BLUPs to test set BLUPs using the kinship.
}
\keyword{internal}
