% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_inv_diag.R
\name{vec.inv.diag}
\alias{vec.inv.diag}
\title{vec.inv.diag}
\usage{
vec.inv.diag(x, y)
}
\arguments{
\item{x}{Something}

\item{y}{Something}
}
\value{
Something
}
\description{
vec.inv.diag
}
\keyword{internal}
