% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRF.R
\name{VRF}
\alias{VRF}
\title{VRF}
\usage{
VRF(sec, train.set, tau = 0.95, verbose = TRUE)
}
\arguments{
\item{sec}{A secondary feature dataframe where the first column contains the genotypes.}

\item{train.set}{A character vector with the names of the training set genotypes.}

\item{tau}{Threshold value. The filtered matrix will have no pair of features with an absolute
correlation larger than this threshold. Defaults to \code{0.95}.}

\item{verbose}{Boolean.}
}
\value{
A list with the VR filtered genetic correlation and covariance matrices, as well as the VR filtered secondary data matrix.
}
\description{
\code{VRF} variance and redundancy filters a secondary feature data matrix. First, features with
negative estimated genetic and/or residual variances are filtered out. Then, the remaining features
are redundancy filtered based on absolute genetic correlations. Note that only training data is used for the filtering.
}
