% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covSS.R
\name{covSS}
\alias{covSS}
\title{covSS}
\usage{
covSS(data, genoMeans = NULL, reps = NULL, sepExp = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{Dataframe containing the genotypes in the first column and features/traits for which (co)variances should
be estimated in the other columns.}

\item{genoMeans}{Genotypic means. \code{NULL} by default and calculated internally. Can be specified to avoid
repeated calculations if covSS is called for the same genotypes multiple times.}

\item{reps}{Number of replicates. Also \code{NULL} by default but can be specified for the same reasons
as \code{genoMeans}.}

\item{sepExp}{Boolean specifying whether the focal trait was measured on different plants than the secondary
features. Only use if there is a focal trait. Sets the residual covariances between secondary features or factors
and the focal to \code{0}.}

\item{verbose}{Boolean}
}
\value{
A list containing the phenotypic, genetic, and residual covariance matrices, as well as a vector
containing the plot level heritabilities.
}
\description{
\code{covSS} calculates the phenotypic, genetic, and residual covariance matrix. If bot \code{sec} and \code{foc}
are specified, individual replicates must be in the same rows in each dataframe!
}
